+++ package.json
@@ -30,7 +30,6 @@
     "*": "pnpm check:fix"
   },
   "dependencies": {
-    "@astro-community/astro-embed-youtube": "^0.5.9",
     "@astrojs/check": "^0.9.5",
     "@astrojs/cloudflare": "12.6.11",
     "@astrojs/markdoc": "^0.15.9",
@@ -52,7 +51,6 @@
     "@sanity/ui": "^3.1.11",
     "@severo_bo/us-atlas-2023": "^4.0.0",
     "astro": "5.16.0",
-    "astro-font": "^1.1.0",
     "astro-github-file-loader": "^1.1.0",
     "astro-portabletext": "^0.12.0",
     "d3": "^7.9.0",
