+++ src/components/react/common/YouTube/YouTube.css
@@ -1,0 +1,3 @@
+.yt-lite:not(.lyt-activated) {
+  mix-blend-mode: luminosity;
+}
