+++ src/components/react/common/YouTube/index.ts
@@ -1,0 +1,1 @@
+export * from "./YouTube";
