+++ src/components/text/YouTube.astro
@@ -1,7 +1,7 @@
 ---
-import { YouTube as YouTubeEmbed } from "@astro-community/astro-embed-youtube";
+import { YouTube as YouTubeComponent } from "@/components/react/common/YouTube";
 
 const { url } = Astro.props.node;
 ---
 
-<YouTubeEmbed id={url} />
+<YouTubeComponent url={url} client:only="react" />
