+++ src/sanity/components/YouTubePreview.tsx
@@ -1,20 +1,14 @@
-import { Flex, Text } from "@sanity/ui";
-import LiteYouTubeEmbed from "react-lite-youtube-embed";
+import { Text } from "@sanity/ui";
+import { YouTube } from "@/components/react/common/YouTube";
 import "react-lite-youtube-embed/dist/LiteYouTubeEmbed.css";
 import type { PreviewProps } from "sanity";
 
 export function YouTubePreview(props: PreviewProps) {
   const { title: url } = props;
 
-  const videoId = typeof url === "string" ? url.split("v=")[1] : null;
-
-  return (
-    <Flex padding={3} align="center" justify="center">
-      {videoId ? (
-        <LiteYouTubeEmbed title="" id={videoId} />
-      ) : (
-        <Text>Add a YouTube URL</Text>
-      )}
-    </Flex>
+  return typeof url === "string" ? (
+    <YouTube url={url} />
+  ) : (
+    <Text>Add a YouTube URL</Text>
   );
 }
